feat(github): improvments according to review
    - move the generation & verification of state in the UI
    - generating github URL in the UI
    - dropped code in the iam related to above 2 points
