+++ app/.env-example.env
@@ -3,6 +3,7 @@ NEXT_PUBLIC_DPOPP_GOOGLE_CLIENT_ID=MY-APP-ID.apps.googleusercontent.com
 NEXT_PUBLIC_DPOPP_TWITTER_CLIENT_ID=ABC123456789
 NEXT_PUBLIC_DPOPP_TWITTER_CALLBACK=http://localhost:3000/
 NEXT_PUBLIC_DPOPP_FACEBOOK_APP_ID=123456789
+NEXT_PUBLIC_DPOPP_GITHUB_CLIENT_ID=12345678
 NEXT_PUBLIC_DPOPP_GITHUB_CALLBACK=http://localhost:3000/
 
 NEXT_PUBLIC_DPOPP_IAM_URL=http://localhost:80/api/
