+++ iam/src/providers/github.ts
@@ -1,9 +1,17 @@
 // ----- Types
 import type { RequestPayload, VerifiedPayload } from "@dpopp/types";
-
-// ----- Github OAuth2
-import { GithubFindMyUserResponse, requestFindMyUser } from "../procedures/githubOauth";
 import type { Provider, ProviderOptions } from "../types";
+import axios from "axios";
+
+export type GithubTokenResponse = {
+  access_token: string;
+};
+
+export type GithubFindMyUserResponse = {
+  id?: string;
+  login?: string;
+  type?: string;
+};
 
 // Export a Github Provider to carry out OAuth and return a record object
 export class GithubProvider implements Provider {
@@ -40,7 +48,40 @@ export class GithubProvider implements Provider {
   }
 }
 
-// Perform verification on twitter access token
-async function verifyGithub(code: string): Promise<GithubFindMyUserResponse> {
-  return await requestFindMyUser(code);
-}
+const requestAccessToken = async (code: string): Promise<string> => {
+  const clientId = process.env.GITHUB_CLIENT_ID;
+  const clientSecret = process.env.GITHUB_CLIENT_SECRET;
+
+  // Exchange the code for an access token
+  const tokenRequest = await axios.post(
+    `https://github.com/login/oauth/access_token?client_id=${clientId}&client_secret=${clientSecret}&code=${code}`,
+    {},
+    {
+      headers: { Accept: "application/json" },
+    }
+  );
+
+  if (tokenRequest.status != 200) {
+    throw `Post for request returned status code ${tokenRequest.status} instead of the expected 200`;
+  }
+
+  const tokenResponse = tokenRequest.data as GithubTokenResponse;
+
+  return tokenResponse.access_token;
+};
+
+const verifyGithub = async (code: string): Promise<GithubFindMyUserResponse> => {
+  // retrieve user's auth bearer token to authenticate client
+  const accessToken = await requestAccessToken(code);
+
+  // Now that we have an access token fetch the user details
+  const userRequest = await axios.get("https://api.github.com/user", {
+    headers: { Authorization: `token ${accessToken}` },
+  });
+
+  if (userRequest.status != 200) {
+    throw `Get user request returned status code ${userRequest.status} instead of the expected 200`;
+  }
+
+  return userRequest.data as GithubFindMyUserResponse;
+};
